chore: split tests by package
