+++ .env.example
@@ -4,3 +4,4 @@ BUNDLER_URL=https://bundler.biconomy.io/api/v2/80002/cJPK7B3ru.dd7f7861-190d-45i
 E2E_BICO_PAYMASTER_KEY_AMOY=
 E2E_BICO_PAYMASTER_KEY_BASE=
 CHAIN_ID=80002
+CODECOV_TOKEN=
