+++ .github/workflows/coverage.yml
@@ -1,7 +1,8 @@
 name: coverage
 on:
-  pull_request:
-    types: [opened, reopened, synchronize, ready_for_review]
+  push:
+    branches:
+      - main
 jobs:
   coverage:
     name: coverage
