+++ tests/account/write.test.ts
@@ -1,18 +1,10 @@
-import {
-  http,
-  createPublicClient,
-  createWalletClient,
-  encodeFunctionData,
-  parseAbi
-} from "viem"
+import { http, createPublicClient, createWalletClient } from "viem"
 import { generatePrivateKey, privateKeyToAccount } from "viem/accounts"
 import { beforeAll, describe, expect, test } from "vitest"
 import {
   type BiconomySmartAccountV2,
-  NATIVE_TOKEN_ALIAS,
   createSmartAccountClient
 } from "../../src/account"
-import { PaymasterMode } from "../../src/paymaster"
 import { testOnlyOnOptimism } from "../setupFiles"
 import { checkBalance, getConfig } from "../utils"
 
@@ -62,73 +54,6 @@ describe("Account: Write", () => {
     )
   })
 
-  test("should send some native token to a recipient", async () => {
-    const balance = (await checkBalance(publicClient, recipient)) as bigint
-    const { wait } = await smartAccount.sendTransaction(
-      {
-        to: recipient,
-        value: 1
-      },
-      {
-        simulationType: "validation_and_execution"
-      }
-    )
-
-    const result = await wait()
-    const newBalance = (await checkBalance(publicClient, recipient)) as bigint
-
-    expect(result?.receipt?.transactionHash).toBeTruthy()
-    expect(newBalance - balance).toBe(1n)
-  }, 50000)
-
-  test("should gaslessly mint an NFT", async () => {
-    const encodedCall = encodeFunctionData({
-      abi: parseAbi(["function safeMint(address to) public"]),
-      functionName: "safeMint",
-      args: [recipient]
-    })
-
-    const transaction = {
-      to: nftAddress, // NFT address
-      data: encodedCall
-    }
-
-    const balance = (await checkBalance(
-      publicClient,
-      recipient,
-      nftAddress
-    )) as bigint
-
-    const maticBalanceBefore = await checkBalance(
-      publicClient,
-      await smartAccount.getAddress()
-    )
-
-    const response = await smartAccount.sendTransaction(transaction, {
-      paymasterServiceData: { mode: PaymasterMode.SPONSORED },
-      simulationType: "validation"
-    })
-
-    const userOpReceipt = await response.wait(3)
-    expect(userOpReceipt.userOpHash).toBeTruthy()
-    expect(userOpReceipt.success).toBe("true")
-
-    const maticBalanceAfter = await checkBalance(
-      publicClient,
-      await smartAccount.getAddress()
-    )
-
-    expect(maticBalanceAfter).toEqual(maticBalanceBefore)
-
-    const newBalance = (await checkBalance(
-      publicClient,
-      recipient,
-      nftAddress
-    )) as bigint
-
-    expect(newBalance - balance).toBe(1n)
-  }, 60000)
-
   test("should deploy a smart account with native token balance", async () => {
     const newPrivateKey = generatePrivateKey()
     const newAccount = privateKeyToAccount(newPrivateKey)
@@ -159,40 +84,6 @@ describe("Account: Write", () => {
 
     // Test:
     const { wait } = await smartAccount.deploy()
-    const { success } = await wait()
-
-    const byteCode = await publicClient.getBytecode({
-      address: smartAccountAddress
-    })
-    expect(success).toBe("true")
-    expect(byteCode).toBeTruthy()
-  }, 60000)
-
-  test("should deploy a smart account with sponsorship", async () => {
-    const newPrivateKey = generatePrivateKey()
-    const newAccount = privateKeyToAccount(newPrivateKey)
-
-    const newViemWallet = createWalletClient({
-      account: newAccount,
-      chain,
-      transport: http()
-    })
-
-    const smartAccount = await createSmartAccountClient({
-      signer: newViemWallet,
-      paymasterUrl,
-      bundlerUrl
-    })
-
-    const smartAccountAddress = await smartAccount.getAccountAddress()
-    const balance = await publicClient.getBalance({
-      address: smartAccountAddress
-    })
-    expect(balance).toBe(0n)
-
-    const { wait } = await smartAccount.deploy({
-      paymasterServiceData: { mode: PaymasterMode.SPONSORED }
-    })
     const { success } = await wait()
 
     const byteCode = await publicClient.getBytecode({
@@ -247,58 +138,6 @@ describe("Account: Write", () => {
     }
   )
 
-  testOnlyOnOptimism(
-    "should gaslessly mint an NFT on optimism",
-    async () => {
-      const encodedCall = encodeFunctionData({
-        abi: parseAbi(["function safeMint(address to) public"]),
-        functionName: "safeMint",
-        args: [recipient]
-      })
-
-      const transaction = {
-        to: nftAddress, // NFT address
-        data: encodedCall
-      }
-
-      const balance = (await checkBalance(
-        publicClient,
-        recipient,
-        nftAddress
-      )) as bigint
-
-      const maticBalanceBefore = await checkBalance(
-        publicClient,
-        await smartAccount.getAddress()
-      )
-
-      const response = await smartAccount.sendTransaction(transaction, {
-        paymasterServiceData: { mode: PaymasterMode.SPONSORED },
-        simulationType: "validation_and_execution"
-      })
-
-      const userOpReceipt = await response.wait(3)
-      expect(userOpReceipt.userOpHash).toBeTruthy()
-      expect(userOpReceipt.success).toBe("true")
-
-      const maticBalanceAfter = await checkBalance(
-        publicClient,
-        await smartAccount.getAddress()
-      )
-
-      expect(maticBalanceAfter).toEqual(maticBalanceBefore)
-
-      const newBalance = (await checkBalance(
-        publicClient,
-        recipient,
-        nftAddress
-      )) as bigint
-
-      expect(newBalance - balance).toBe(1n)
-    },
-    50000
-  )
-
   test.skip("should withdraw erc20 balances", async () => {
     const usdt = "0xda5289fcaaf71d52a80a254da614a192b693e977"
     const smartAccountOwner = walletClient.account.address
@@ -342,181 +181,6 @@ describe("Account: Write", () => {
     expect(usdtBalanceOfSAAfter - usdtBalanceOfSABefore).toBe(-1n)
     expect(usdtBalanceOfRecipientAfter - usdtBalanceOfRecipientBefore).toBe(1n)
   }, 25000)
-
-  test.skip("should gaslessly withdraw nativeToken", async () => {
-    const smartAccountOwner = walletClient.account.address
-
-    const smartAccountAddress = await smartAccount.getAddress()
-    const balanceOfSABefore = (await checkBalance(
-      publicClient,
-      smartAccountAddress
-    )) as bigint
-    const balanceOfRecipientBefore = (await checkBalance(
-      publicClient,
-      smartAccountOwner
-    )) as bigint
-
-    const { wait } = await smartAccount.withdraw(
-      [
-        {
-          address: NATIVE_TOKEN_ALIAS,
-          amount: BigInt(1),
-          recipient: smartAccountAddress
-        }
-      ],
-      null,
-      {
-        paymasterServiceData: { mode: PaymasterMode.SPONSORED }
-      }
-    )
-
-    const {
-      receipt: { transactionHash },
-      userOpHash,
-      success
-    } = await wait()
-
-    expect(userOpHash).toBeTruthy()
-    expect(success).toBe("true")
-    expect(transactionHash).toBeTruthy()
-
-    const balanceOfSAAfter = (await checkBalance(
-      publicClient,
-      smartAccountAddress
-    )) as bigint
-    const balanceOfRecipientAfter = (await checkBalance(
-      publicClient,
-      smartAccountOwner
-    )) as bigint
-
-    expect(balanceOfSABefore - balanceOfSAAfter).toBe(1n)
-    expect(balanceOfRecipientAfter - balanceOfRecipientBefore).toBe(1n)
-  }, 25000)
-
-  test("should withdraw nativeToken and an erc20 token", async () => {
-    const usdt = "0xda5289fcaaf71d52a80a254da614a192b693e977"
-    const smartAccountOwner = walletClient.account.address
-
-    const smartAccountAddress = await smartAccount.getAddress()
-    const balanceOfSABefore = (await checkBalance(
-      publicClient,
-      smartAccountAddress
-    )) as bigint
-    const balanceOfRecipientBefore = (await checkBalance(
-      publicClient,
-      smartAccountOwner
-    )) as bigint
-    const usdtBalanceOfSABefore = await checkBalance(
-      publicClient,
-      smartAccountAddress,
-      usdt
-    )
-    const usdtBalanceOfRecipientBefore = await checkBalance(
-      publicClient,
-      smartAccountOwner,
-      usdt
-    )
-
-    const { wait } = await smartAccount.withdraw(
-      [
-        { address: usdt, amount: BigInt(1) },
-        { address: NATIVE_TOKEN_ALIAS, amount: BigInt(1) }
-      ],
-      smartAccountOwner,
-      {
-        paymasterServiceData: { mode: PaymasterMode.SPONSORED }
-      }
-    )
-
-    const {
-      receipt: { transactionHash },
-      userOpHash,
-      success
-    } = await wait()
-
-    expect(userOpHash).toBeTruthy()
-    expect(success).toBe("true")
-    expect(transactionHash).toBeTruthy()
-
-    const balanceOfSAAfter = (await checkBalance(
-      publicClient,
-      smartAccountAddress
-    )) as bigint
-    const balanceOfRecipientAfter = (await checkBalance(
-      publicClient,
-      smartAccountOwner
-    )) as bigint
-    const usdtBalanceOfSAAfter = (await checkBalance(
-      publicClient,
-      smartAccountAddress,
-      usdt
-    )) as bigint
-    const usdtBalanceOfRecipientAfter = (await checkBalance(
-      publicClient,
-      smartAccountOwner,
-      usdt
-    )) as bigint
-
-    expect(balanceOfSABefore - balanceOfSAAfter).toBe(1n)
-    expect(balanceOfRecipientAfter - balanceOfRecipientBefore).toBe(1n)
-
-    expect(usdtBalanceOfSAAfter - usdtBalanceOfSABefore).toBe(-1n)
-    expect(usdtBalanceOfRecipientAfter - usdtBalanceOfRecipientBefore).toBe(1n)
-  }, 60000)
-
-  test("should withdraw all native token", async () => {
-    const smartAccountOwner = walletClient.account.address
-    const smartAccountAddress = await smartAccount.getAddress()
-    const balanceOfSABefore = (await checkBalance(
-      publicClient,
-      smartAccountAddress
-    )) as bigint
-    const balanceOfRecipientBefore = (await checkBalance(
-      publicClient,
-      smartAccountOwner
-    )) as bigint
-
-    const { wait } = await smartAccount.withdraw(
-      [] /* null or undefined or [] */,
-      smartAccountOwner,
-      {
-        paymasterServiceData: { mode: PaymasterMode.SPONSORED } // Will leave no dust
-      }
-    )
-
-    const {
-      receipt: { transactionHash },
-      userOpHash,
-      success
-    } = await wait()
-
-    expect(userOpHash).toBeTruthy()
-    expect(success).toBe("true")
-    expect(transactionHash).toBeTruthy()
-
-    const balanceOfSAAfter = (await checkBalance(
-      publicClient,
-      smartAccountAddress
-    )) as bigint
-    const balanceOfRecipientAfter = (await checkBalance(
-      publicClient,
-      smartAccountOwner
-    )) as bigint
-
-    expect(balanceOfSAAfter).toBe(0n)
-    expect(balanceOfRecipientAfter).toBe(
-      balanceOfSABefore + balanceOfRecipientBefore
-    )
-
-    // Teardown: send back the native token to the smart account
-    const teardownHash = await walletClient.sendTransaction({
-      to: smartAccountAddress,
-      value: balanceOfSABefore,
-      account,
-      chain
-    })
-    expect(teardownHash).toBeTruthy()
-  }, 60000)
 
   test.skip("should mint an NFT on Mumbai and pay with ERC20 - with preferredToken", async () => {
     // const {
