+++ tests/bundler/read.test.ts
@@ -1,0 +1,168 @@
+import { http, type Chain, createWalletClient } from "viem"
+import { privateKeyToAccount } from "viem/accounts"
+import { beforeAll, describe, expect, test } from "vitest"
+import {
+  type BiconomySmartAccountV2,
+  type BiconomySmartAccountV2Config,
+  createSmartAccountClient,
+  compareChainIds
+} from "../../src/account"
+import { createBundler } from "../../src/bundler"
+import { getBundlerUrl, getConfig } from "../utils"
+
+describe("Bundler: Read", () => {
+  const {
+    chain,
+    chainId,
+    privateKey,
+    privateKeyTwo,
+    bundlerUrl,
+    paymasterUrl
+  } = getConfig()
+  const account = privateKeyToAccount(`0x${privateKey}`)
+  const accountTwo = privateKeyToAccount(`0x${privateKeyTwo}`)
+  let [smartAccount, smartAccountTwo]: BiconomySmartAccountV2[] = []
+
+  const [walletClient, walletClientTwo] = [
+    createWalletClient({
+      account,
+      chain,
+      transport: http()
+    }),
+    createWalletClient({
+      account: accountTwo,
+      chain,
+      transport: http()
+    })
+  ]
+
+  beforeAll(async () => {
+    ;[smartAccount, smartAccountTwo] = await Promise.all(
+      [walletClient, walletClientTwo].map((client) =>
+        createSmartAccountClient({
+          chainId,
+          signer: client,
+          bundlerUrl,
+          paymasterUrl
+        })
+      )
+    )
+  })
+
+  test.concurrent(
+    "should parse the rpcUrl when a custom chain and bundler are used",
+    async () => {
+      const customBlastChain = {
+        id: 81_457,
+        name: "Blast",
+        //   network: "blast",
+        nativeCurrency: {
+          decimals: 18,
+          name: "Ethereum",
+          symbol: "ETH"
+        },
+        rpcUrls: {
+          public: { http: ["https://rpc.blast.io"] },
+          default: { http: ["https://rpc.blast.io"] }
+        },
+        blockExplorers: {
+          etherscan: { name: "Blastscan", url: "https://blastscan.io/" },
+          default: { name: "Blastscan", url: "https://blastscan.io/" }
+        },
+        contracts: {
+          multicall3: {
+            address: "0xca11bde05977b3631167028862be2a173976ca11",
+            blockCreated: 88_189
+          }
+        }
+      } as const satisfies Chain
+
+      const accountOne = privateKeyToAccount(`0x${privateKey}`)
+
+      const walletClientWithCustomChain = createWalletClient({
+        account: accountOne,
+        chain: customBlastChain,
+        transport: http()
+      })
+
+      const blastBundler = await createBundler({
+        bundlerUrl: getBundlerUrl(customBlastChain.id),
+        viemChain: customBlastChain
+      })
+      const smartAccountFromViemWithCustomChain =
+        await createSmartAccountClient({
+          viemChain: customBlastChain,
+          signer: walletClientWithCustomChain,
+          bundler: blastBundler,
+          rpcUrl: customBlastChain.rpcUrls.default.http[0]
+        })
+
+      expect(
+        smartAccountFromViemWithCustomChain.rpcProvider.transport.url
+      ).toBe("https://rpc.blast.io")
+      expect(blastBundler.getBundlerUrl()).toBe(
+        getBundlerUrl(customBlastChain.id)
+      )
+    }
+  )
+
+  test.concurrent(
+    "should throw an error, bundlerUrl chain id and signer chain id does not match",
+    async () => {
+      const config: BiconomySmartAccountV2Config = {
+        signer: walletClient,
+        bundlerUrl: getBundlerUrl(1337),
+        paymasterUrl
+      }
+
+      await expect(
+        compareChainIds(walletClient, config, false)
+      ).rejects.toThrow()
+    }
+  )
+  test.concurrent(
+    "should throw an error, bundlerUrl chainId and paymasterUrl + chainId does not match",
+    async () => {
+      const mockPaymasterUrl =
+        "https://paymaster.biconomy.io/api/v1/1337/-RObQRX9ei.fc6918eb-c582-4417-9d5a-0507b17cfe71"
+
+      const config: BiconomySmartAccountV2Config = {
+        signer: walletClient,
+        bundlerUrl,
+        paymasterUrl: mockPaymasterUrl
+      }
+
+      await expect(
+        compareChainIds(walletClient, config, false)
+      ).rejects.toThrow()
+    }
+  )
+
+  test.concurrent(
+    "should throw, chain id from signer and bundlerUrl do not match",
+    async () => {
+      const createAccount = createSmartAccountClient({
+        signer: walletClient,
+        bundlerUrl:
+          "https://bundler.biconomy.io/api/v2/1/nJPK7B3ru.dd7f7861-190d-41bd-af80-6877f74b8f44" // mock
+      })
+
+      await expect(createAccount).rejects.toThrow()
+    }
+  )
+
+  test.concurrent(
+    "should throw, chain id from paymasterUrl and bundlerUrl do not match",
+    async () => {
+      const createAccount = createSmartAccountClient({
+        signer: walletClient,
+        paymasterUrl:
+          "https://paymaster.biconomy.io/api/v1/1/-RObQRX9ei.fc6918eb-c582-4417-9d5a-0507b17cfe71",
+        bundlerUrl:
+          "https://bundler.biconomy.io/api/v2/80002/nJPK7B3ru.dd7f7861-190d-41bd-af80-6877f74b8f44" // mock
+      })
+
+      await expect(createAccount).rejects.toThrow()
+    }
+  )
+})
